Fix: Finally eliminate ajv error & Log url to user on deploy (#76)
We move the ajv dependencies out of dev deps. This actually resolves the
issue.

running make-agent deploy without the [optional] url argument results in
the following:

```
➜  wraptor git:(upgrade) ✗ npx make-agent deploy
Retrying...
Retrying...
Retrying...
Unexpected error: fetch failed
OpenAPI specification validation failed.
```

Digging into the code it seems that the env vars are not loaded. There
is also no documentation about what env vars should be set if the URL is
not supplied in the runtime argument.


Closes #77